--- a/src/renderer/mod.rs
+++ b/src/renderer/mod.rs
@@ -13,6 +13,8 @@ pub mod renderer {
     use sdl2::pixels::PixelFormatEnum;
     use std::sync::{Arc, Mutex};
     use num;
+    use time::PreciseTime;
+
 
     #[derive(Clone, Debug)]
     pub enum RendererError {
@@ -32,6 +34,7 @@ pub mod renderer {
         viewport_width: u32,
         viewport_height: u32,
         viewport_offset: Point,
+        frame: u32,
     }
 
     impl Renderer {
@@ -81,6 +84,7 @@ pub mod renderer {
                    viewport_width: 0,
                    viewport_height: 0,
                    viewport_offset: Point::new(0, 0),
+                   frame: 0,
                })
         }
 
@@ -88,13 +92,45 @@ pub mod renderer {
             if self.viewport_width == 0 {
                 self.update_dimensions();
             }
+
+            // Translate the pixel values to RGB colors.
+            let src_buffer = screen.lock().unwrap().back_buffer;
+            let mut rgb_buffer = *screen.lock().unwrap().buffer_rgb;
+            let mut palette = px8::PALETTE.lock().unwrap();
+
+            let mut j = 0;
+            let mut cached_pixel: u8 = 0;
+            let mut rgb = palette.get_rgb(cached_pixel as u32);
+
+            let start = PreciseTime::now();
+
+            for pixel in src_buffer.iter() {
+                if *pixel != cached_pixel {
+                    rgb = palette.get_rgb(*pixel as u32);
+                    cached_pixel = *pixel;
+                }
+                rgb_buffer[j] = rgb.b;
+                rgb_buffer[j + 1] = rgb.g;
+                rgb_buffer[j + 2] = rgb.r;
+                j = j + 3;
+            }
+
+            let t1 = PreciseTime::now();
+
+            // Update the texture with the RGB values.
             self.texture
-                .update(None,
-                        &mut *screen.lock().unwrap().buffer_rgb,
-                        px8::SCREEN_WIDTH * 3)
+                .update(None, &mut rgb_buffer, px8::SCREEN_WIDTH * 3)
                 .unwrap();
 
-            self.renderer.clear();
+            let t2 = PreciseTime::now();
+
+            // Copy the texture to the screen.
+            // Only need to clear the screen if there is any border -
+            // the main display area will be overwritten.
+            if self.viewport_offset.x() != 0 || self.viewport_offset.y() != 0 {
+                self.renderer.clear();
+            }
+
             self.renderer
                 .copy(&self.texture,
                       Some(Rect::new(0, 0, px8::SCREEN_WIDTH as u32, px8::SCREEN_HEIGHT as u32)),
@@ -103,7 +139,24 @@ pub mod renderer {
                                      self.viewport_width,
                                      self.viewport_height)))
                 .unwrap();
+
+            let t3 = PreciseTime::now();
+
             self.renderer.present();
+
+            let t4 = PreciseTime::now();
+
+            if cfg!(feature = "blit_perf") {
+                if self.frame % 60 == 0 {
+                    println!("gen_rgb:{} update_tex:{} copy_tex:{} present:{}",
+                             start.to(t1),
+                             t1.to(t2),
+                             t2.to(t3),
+                             t3.to(t4))
+                }
+            }
+
+            self.frame += 1;
         }
 
         pub fn update_dimensions(&mut self) {