@hallucino 2017-05-23
I did some tests on raspberry pi and it works really correctly