@hallucino 2017-05-26
I added some checks that you removed because I got some crashes:
https://github.com/Gigoteur/PX8/commit/7f34f5859c0cd111f2baa52b528aa52d5cbca43d